requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pymilvus==2.4.9
google-generativeai>=0.8.0
pandas==2.2.0
openpyxl==3.1.2
langdetect==1.0.9
python-dotenv==1.0.0
python-multipart==0.0.9
aiofiles==23.2.1
